Move filtering higher in the chain
